backend/back/requirement.txt
adblockparser==0.7
AdvancedHTTPServer==2.2.0
aiodns==2.0.0
aiohttp==3.6.2
aioredis==1.3.1
ajpy==0.0.4
alembic==1.1.0.dev0
aniso8601==8.0.0
APScheduler==0.0.0
asn1crypto==0.24.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.8.0
backcall==0.1.0
backdoor-factory==0.0.0
basemap==1.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
binwalk==2.2.0
bleach==3.1.5
blinker==1.4
boltons==19.1.0
bottle==0.12.15
Brotli==1.0.7
cairocffi==0.9.0
capstone==3.0.5
catfish==1.4.13
cbor==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
Chameleon==3.6.2
chardet==3.0.4
CherryPy==8.9.1
chrome-gnome-shell==0.0.0
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crackmapexec==5.0.1.dev0
crcelk==1.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debtags==2.1
decorator==4.4.2
defusedxml==0.6.0
deluge==2.0.3
deprecation==2.0.7
dhcpig==0.0.0
dicttoxml==1.7.4
distro==1.5.0
distro-info==0.23
Django==2.2.13
django-allauth==0.42.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.3.0
django-filters==0.2.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.0.7
djangorestframework==3.11.0
dnslib==0.9.13
dnspython==1.16.0
dynaconf==2.2.3
ecdsa==0.15
email-validator==1.0.3
entrypoints==0.3
et-xmlfile==1.0.1
ExifRead==2.1.2
faraday-plugins==1.2
faradaysec==3.10.2
feedparser==5.2.1
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.4
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-KVSession-fork==0.6.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Restless==0.17.0
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.18.2
GDAL==3.0.4
GeoIP==1.3.2
geoip2==2.9.0
geojson==2.5.0
gevent==1.4.0
gitdb==4.0.5
GitPython==3.1.3
gpg===1.13.1-unknown
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.15
grequests==0.4.0
gyp==0.1
h11==0.9.0
h2==3.2.0
hashID==3.1.4
hiredis==1.0.1
hpack==3.0.0
html2text==2020.1.16
html5lib==1.0.1
hupper==1.10.2
hvac==0.10.3
hyperframe==5.2.0
hyperlink==19.0.0
icalendar==4.0.3
idna==2.9
impacket==0.9.21
importlib-metadata==1.6.0
incremental==16.10.1
ipwhois==0.15.1
IPy==1.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.0
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kaitaistruct==0.8
kazam==1.4.5
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2019.9.1
KismetCaptureRtladsb==2019.10.1
KismetCaptureRtlamr==2019.10.1
kiwisolver==1.0.1
knox==0.1.0
lazy-object-proxy==1.4.3
ldap3==2.7
ldapdomaindump==0.9.1
lightdm-gtk-greeter-settings==1.2.2
loguru==0.5.1
lxml==4.5.0
lz4==3.0.2+dfsg
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.19.0
matplotlib==3.2.1
maxminddb==1.4.1
mccabe==0.6.1
mechanize==0.4.5
mimerender==0.6.0
mistune==0.8.4
mitmproxy==4.0.4
more-itertools==4.2.0
msgpack==0.6.2
multidict==4.7.6
mysqlclient==1.4.4
nassl==3.0.0
nbconvert==5.6.1
nbformat==5.0.7
netaddr==0.7.19
networkx==2.4
nodejs==0.1.1
notebook==6.0.3
nplusone==1.0.0
npm==0.1.1
ntlm-auth==1.4.0
numexpr==2.7.1
numpy==1.17.4
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==2.4.9
optional-django==0.1.0
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.0
passlib==1.7.2
Paste==3.4.0
PasteDeploy==2.1.0
PasteScript==2.0.2
patator==0.7
pefile==2019.4.18
pexpect==4.6.0
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==6.2.1
plaster==1.0
plaster-pastedeploy==0.5
plecost==1.1.2
plotly==4.7.1
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
prometheus-client==0.8.0
promise==2.2
prompt-toolkit==3.0.5
protobuf==3.11.4
psycopg2==2.8.5
ptyprocess==0.6.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycares==3.1.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==1.9.73
pycurl==7.43.0.2
pydot==1.4.1
pyenchant==2.0.0
pyExploitDb==0.2.0
pyfiglet==0.8.post0
Pygments==2.3.1
PyGObject==3.36.0
pygraphviz==1.5
PyHamcrest==1.9.0
pyinotify==0.9.6
pylint==2.5.3
pylnk3==0.2.1
pymssql==2.1.4
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.7
pypcapfile==0.12.1
PyPDF2==1.26.0
pyperclip==1.8.0
pypng==0.0.20
pyproj==2.6.1
PyQRCode==1.2.1
PyQt5==5.14.2
pyqtgraph==0.11.0rc0
pyramid==1.10.4
pyrsistent==0.15.5
pyserial==3.4
pyShodan==0.2.3
pyshp==2.1.0
pysmbc==1.0.15.6
pysmi==0.3.2
pysnmp==4.4.6
PySocks==1.6.8
python-apt==2.1.3
python-box==3.4.6
python-dateutil==2.8.1
python-debian==0.1.37
python-docx==0.8.10
python-dotenv==0.13.0
python-editor==1.0.3
python-libtorrent==1.1.13
python-magic==0.4.16
python-magic-ahupp==0.4.13
python-mimeparse==1.6.0
python-pam==1.8.4
python-pptx==0.6.18
python-snappy==0.5.3
python3-openid==3.1.0
PyTrie==0.2
pytz==2020.1
pywerview==0.2.0
pywinrm==0.4.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
qrcode==6.1
Quamash==0.6.1
redis==3.3.11
rencode==1.0.6
repoze.lru==0.7
requests==2.23.0
requests-file==1.5.1
requests-futures==1.0.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retrying==1.3.3
Routes==2.4.1
rq==1.4.0
ruamel.yaml==0.15.89
rule-engine==1.1.0
Rx==1.6.1
scapy==2.4.3
scipy==1.4.1
secure==0.2.1
selenium==4.0.0a1
Send2Trash==1.5.0
service-identity==18.1.0
setproctitle==1.1.10
shodan==1.23.0
shortuuid==1.0.1
simplejson==3.17.0
simplekv==0.13.0
sip==4.19.22
six==1.9.0
smmap==3.0.4
smoke-zephyr==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.15
sqlalchemy-schemadisplay==1.3
sqlparse==0.3.1
sslyze==3.0.6
stem==1.7.1
syslog-rfc5424-formatter==1.2.2
tables==3.6.1
tabulate==0.8.2
Tempita==0.5.2
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
termineter==1.0.4
testpath==0.4.4
texttable==1.6.2
theHarvester==3.1.0
tld==0.11.11
tls-parser==1.2.2
toml==0.10.1
tornado==5.1.1
tqdm==4.43.0
traitlets==4.3.3
translationstring==1.3
Twisted==18.9.0
txaio==20.4.1
typing-extensions==3.7.4.2
tzlocal==2.1b1
u-msgpack-python==2.3.0
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
urwid==2.0.1
venusian==3.0.0
vinetto==0.8.0
wafw00f==2.1.0
waitress==1.4.1
wapiti3==3.0.3
wcwidth==0.1.9
webargs==6.1.0
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.53.0
WebTest==2.0.34
Werkzeug==1.0.0
wfuzz==2.4.5
whois==0.8
wifite==2.5.2
wrapt==1.12.1
wsaccel==0.6.2
wsproto==0.15.0
WTForms==2.2.1
xcffib==0.8.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmltodict==0.12.0
yara-python==3.10.0
yarl==1.4.2
yaswfp==0.9.3
zim==0.72.1
zipp==1.0.0
zope.component==4.3.0
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==4.7.1
